requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.6.1
asyncprawcore==2.3.0
attrs==22.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.3
cymem==2.0.7
dadjokes==1.3.2
discord==1.7.3
discord.py==1.7.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
gazpacho==1.1
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.1
multidict==6.0.4
murmurhash==1.0.9
newsapi-python==0.2.6
numpy==1.21.6
obliquestrategies==0.3
packaging==22.0
pathy==0.10.1
preshed==3.0.8
pydantic==1.10.2
quote==2.0.4
redis==4.4.1
requests==2.28.1
six==1.16.0
smart-open==6.3.0
spacy==3.4.4
spacy-legacy==3.0.11
spacy-loggers==1.0.4
spotipy==2.22.0
srsly==2.4.5
thinc==8.1.6
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.1.1
update-checker==0.18.0
urllib3==1.26.13
wasabi==0.10.1
yarl==1.8.2
zipp==3.11.0
